Mim2GeneParser: add direct rather than dependent xrefs wherever possible
If an entry has got an Ensembl stable ID, add a direct xref regardless
of whether or not it has got an EntrezGene ID. Dependent EntrezGene xrefs
are now only added for entries without Ensembl ID (but with an EntrezGene
one, of course).

WARNING: for the time being this is SERIOUSLY ugly code because even though
direct and dependent xref share a lot of logic, the code handling it is
duplicated. Will try to improve this soon.