Mim2GeneParser: is_header_file_valid() now parses header fields iteratively
This way we avoid having array indices hard-coded in + the function should
be easier to extend if necessary. Note that this version only works if we do not
capture anything from the headers (which should be easy enough to implement if
need be) and that each field is completely independent from the rest (which
would be trickier).