BaseParser: factor dependent-xref creation out of add_dependent_xref()
Functionally, creation of dependent xrefs should now be the same as for direct ones:
there is a heavyweight function for doing everything plus a lightweight one for
mappings only. Unfortunately respective method names are not the same because of a
conflict with existing dependent-xref methods.

API consistency aside, there were two reasons for having done this:
 - there are some, or perhaps quite a few, cases when we do NOT want to create stub
   xrefs before establishing a mapping. For example, Mim2Parser has not got enough
   information to create proper MIM xrefs and should strictly rely on MIMParser ones
   instead of trying to create its own;
 - like add_to_direct_xrefs(), add_dependent_xref() - or rather, the method get_xref()
   that they both call - does not support the possibility of having multiple xref_ids
   matching a single accession. Not quite sure how such a thing would work in the
   first place, however certain parsers (e.g. Mim2GeneParser again) do consider such
   a possibility - so there you go.