BaseParser: have add_to_direct_xref() actually use linkage
Before, the method would accept the parameter 'linkage' but then happily
do squat all with it and pass an empty string to the relevant field of SQL
insert. Now we do pass $linkage to the insert.

As an added bonus, if 'linkage' is not set the relevant field is set to NULL
rather than an empty string.