BaseParser::get_source_id_for_source_name(): throw on missing source
Returning -1 in the event of there being no source ID matching the given
name/priority is kind of pointless, if a parser asks for the ID of a
source chances are pretty good that it actually wants to use it. Indeed,
every single parser I have seen so far implemented its own "if any
retrieved source ID is -1, abort" logic. The warning message with its
condemning "the source name is harcoded in the parser" didn't help
either, seeing that

a) it was just an educated guess because get_source_id_for_source_name()
   didn't actually look at the code of the calling function, and
b) there are perfectly valid reasons for having hardcoded source names,
   for instance when a parser adds xrefs associated with many sources.