MGI_Desc_Parser: fix the handling of " in synonyms
Fields in MRK_List2.rpt are never quoted in any way but as of October
2019 there are six occurrences of the character " - which, just to make
it clear, are part of respective synonyms - in the "list of synonyms"
field:
 - "VC"
 - hLCAT-tg subline A31 ("U line" hemizygous 120 copy)
 - hLCAT-tg subline T59 ("Z line" hemizygous 15 copy)

By default Text::CSV uses " as both the quote and the escape character so
it quietly discarded the quotation marks in the former and complained about
loose quotes (by which it means quotes appearing in an unquoted field)
upon encountering the first of the latter two. Prem seems to have missed
the former and only enabled allow_loose_quotes in the parser to let the
latter through.

The proper way of doing this is of course to tell Text::CSV not to
assign any special meaning to quotation marks, which is what the commit
at hand attempts to do. No more output discrepancies from the output of
the pre-sprint version of the parser, and no need to allow loose quotes
either.